Assignment One
This is assignment one: to print a formatted calendar in the terminal.
It is not fully functional as the some of the lines are getting messed
up, and the string concatenation is being weird shifting months around.
Furthermore, I am having issues with the months like February where the
for loop goes past the explicit 28 days.